[MOVED] Open weather service and model to separate module
